+++ src/Core/Domain/Alias/QueryHandler/SearchForSearchTermHandlerInterface.php
@@ -24,24 +24,20 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-declare(strict_types=1);
+namespace PrestaShop\PrestaShop\Core\Domain\Alias\QueryHandler;
 
-namespace PrestaShop\PrestaShop\Core\Domain\Alias\Query;
-
-use PrestaShop\PrestaShop\Core\Exception\InvalidArgumentException;
+use PrestaShop\PrestaShop\Core\Domain\Alias\Query\SearchForSearchTerm;
+use PrestaShop\PrestaShop\Core\Domain\Alias\QueryResult\SearchTerm;
 
 /**
- * Class SearchAliasesForAssociation is responsible for searching aliases with particular search terms.
+ * Interface SearchAliasesForAssociationHandlerInterface defines contract for SearchAliasesForAssociationHandler
  */
-class SearchAliasesForAssociation
+interface SearchForSearchTermHandlerInterface
 {
     /**
-     * @throws InvalidArgumentException
+     * @param SearchForSearchTerm $query
+     *
+     * @return SearchTerm[]
      */
-    public function __construct(public readonly string $searchTerm, public readonly ?int $limit = null)
-    {
-        if (null !== $limit && $limit <= 0) {
-            throw new InvalidArgumentException('Search limit must be a positive integer or null');
-        }
-    }
+    public function handle(SearchForSearchTerm $query): array;
 }
